Migrate meet the team page so we have a real page to test
